Pre-build update
Added UI functions
Added Enemy AI
Added object behavior

+++ Assets/Scripts/ObjectController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ObjectController : MonoBehaviour
 {
-    public float energyAmount;
+    public int energyAmount;
     PlayerController player;
 
     private void Start()
